requirements.txt
numpy
opencv-python-headless
scikit-image
matplotlib
streamlit
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
pycocotools
Pillow
torch==1.8.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html